break controller and input classes out to their own files
