+++ package/batocera/core/batocera-configgen/configgen/configgen/generators/flycast/flycastControllers.py
@@ -8,7 +8,7 @@ from ...batoceraPaths import mkdir_if_not_exists
 from .flycastPaths import FLYCAST_MAPPING
 
 if TYPE_CHECKING:
-    from ...controllersConfig import Controller
+    from ...controller import Controller
 
 eslog = logging.getLogger(__name__)
