+++ package/batocera/core/batocera-configgen/configgen/configgen/generators/fsuae/fsuaeControllers.py
@@ -6,7 +6,7 @@ from ...batoceraPaths import mkdir_if_not_exists
 from .fsuaePaths import FSUAE_CONFIG_DIR
 
 if TYPE_CHECKING:
-    from ...controllersConfig import ControllerMapping
+    from ...controller import ControllerMapping
     from ...Emulator import Emulator
